+++ app/controllers/api/v1/articles_controller.rb
@@ -1,0 +1,55 @@
+class Api::V1::ArticlesController < ApplicationController
+  before_action :set_article, only: %i[show update destroy]
+
+  # GET /articles
+  # GET /articles.json
+  def index
+    @articles = Article.all
+    render json: @articles, each_serializer: ArticleSerializer
+  end
+
+  # GET /articles/1
+  # GET /articles/1.json
+  def show
+  end
+
+  # POST /articles
+  # POST /articles.json
+  def create
+    @article = Article.new(article_params)
+
+    if @article.save
+      render json: @article, status: :created
+    else
+      render json: @article.errors, status: :unprocessable_entity
+    end
+  end
+
+  # PATCH/PUT /articles/1
+  # PATCH/PUT /articles/1.json
+  def update
+    if @article.update(article_params)
+      render json: @article, status: :ok
+    else
+      render json: @article.errors, status: :unprocessable_entity
+    end
+  end
+
+  # DELETE /articles/1
+  # DELETE /articles/1.json
+  def destroy
+    @article.destroy
+  end
+
+  private
+
+    # Use callbacks to share common setup or constraints between actions.
+    def set_article
+      @article = Article.find(params[:id])
+    end
+
+    # Only allow a list of trusted parameters through.
+    def article_params
+      params.require(:article).permit(:title, :body)
+    end
+end
