+++ app/models/article.rb
@@ -3,5 +3,5 @@ class Article < ApplicationRecord
   belongs_to :user
   has_many :comments, dependent: :destroy
   has_many :article_likes, dependent: :destroy
-  validates :title,  presence: true
+  validates :title, presence: true
 end
