+++ app/serializers/article_serializer.rb
@@ -1,3 +1,3 @@
 class ArticleSerializer < ActiveModel::Serializer
-   attributes :id, :title, :body
+  attributes :id, :title, :body
 end
