+++ spec/models/user_spec.rb
@@ -1,7 +1,6 @@
 require "rails_helper"
 
 RSpec.describe User, type: :model do
-
   context "必要な情報が揃っている場合" do
     let(:user) { build(:user) }
 
@@ -9,8 +8,6 @@ RSpec.describe User, type: :model do
       expect(user).to be_valid
     end
   end
-
-
 
   context "名前のみ入力している場合" do
     let(:user) { build(:user, email: nil, password: nil) }
