+++ spec/requests/api/v1/base_api_spec.rb
@@ -1,4 +1,4 @@
-require 'rails_helper'
+require "rails_helper"
 
 RSpec.describe "Api::V1::BaseApis", type: :request do
   describe "GET /index" do
